\name{align.plots}
\alias{align.plots}
\title{Adapted from ggExtra package which is no longer available. This is related to
an experimental mlpsa plot that will combine the circular plot along with
the two individual distributions.}
\usage{
  align.plots(gl, ...)
}
\description{
  Adapted from ggExtra package which is no longer
  available. This is related to an experimental mlpsa plot
  that will combine the circular plot along with the two
  individual distributions.
}

