% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.balance.R
\name{as.data.frame.covariate.balance}
\alias{as.data.frame.covariate.balance}
\title{Returns the overall effects as a data frame.}
\usage{
\method{as.data.frame}{covariate.balance}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{results of [covariate.balance()].}

\item{row.names}{unused.}

\item{optional}{unused.}

\item{...}{unused}
}
\value{
a data frame with overall covariate effects before and after adjustment.
}
\description{
Returns the overall effects as a data frame.
}
