\name{covariate.balance}
\alias{covariate.balance}
\title{Estimate covariate effect sizes before and after propensity score adjustment.}
\usage{
  covariate.balance(covs, treat, level2, strata)
}
\arguments{
  \item{covs}{frame or matrix of covariates.}

  \item{treat}{vector of treatment indicators.}

  \item{level2}{vector indicating level 2 membership.}

  \item{strata}{strata indicators.}
}
\description{
  Estimate covariate effect sizes before and after
  propensity score adjustment.
}

