% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.trans.psa.R
\name{cv.trans.psa}
\alias{cv.trans.psa}
\title{Transformation of Factors to Individual Levels}
\usage{
cv.trans.psa(covariates, fcol = NULL)
}
\arguments{
\item{covariates}{A dataframe of covariates, presumably some factors.}

\item{fcol}{An optional vector containing the factor columns in the covariate 
data frame. In \code{NULL} (default) routine to identify factors internally.}
}
\description{
The function \code{cv.trans.psa} takes a covariate data frame and replaces each 
categorical covariate of \code{n >=3} levels with n new binary covariate columns, 
one for each level. Transforms covariate dataframe for use with the function 
\code{cv.bal.psa}.
}
\details{
NOTE: This function originated in the \code{PSAgraphics} package. It has been
adapted here for the \code{multilevelPSA} package.
}
\author{
James E. Helmreich James.Helmreich@Marist.edu

Robert M. Pruzek RMPruzek@yahoo.com

KuangNan Xiong harryxkn@yahoo.com

Jason Bryer jason@bryer.org
}
