\name{geom_rug_alt}
\alias{geom_rug_alt}
\title{A ggplot2 alternate rug geomtry.}
\usage{
  geom_rug_alt(mapping = NULL, data = NULL,
    stat = "identity", position = "identity", ...)
}
\arguments{
  \item{mapping}{see ggplot2}

  \item{data}{see ggplot2}

  \item{stat}{see ggplot2}

  \item{position}{see ggplot2}

  \item{...}{other parameters passed to ggplot2}
}
\description{
  This creates an additional ggplot2 geometry
  \code{geom_rug_alt} which allows for rug plots to be
  placed on the top and right of the graph.
}
\author{
  Jason Bryer \email{jason@bryer.org}

  William Doane
}
\references{
  \url{http://stackoverflow.com/questions/4867597/can-you-easily-plot-rugs-axes-on-the-top-right-in-ggplot2}
}

