\docType{data}
\name{GeomRugAlt}
\alias{GeomRugAlt}
\title{A ggplot2 alternate rug geomtry.}
\format{proto object 
 $ draw        :function (., data, scales, coordinates, ...)   
 $ default_stat:function (.)   
 $ objname     : chr "rug_alt" 
 $ default_aes :function (.)   
 $ examples    :function (.)   
 $ desc        : chr "Marginal rug plots" 
 $ guide_geom  :function (.)   
 parent: proto object 
 .. parent: proto object}
\usage{
  GeomRugAlt
}
\description{
  This creates an additional ggplot2 geometry
  \code{geom_rug_alt} which allows for rug plots to be
  placed on the top and right of the graph.
}
\seealso{
  geom_rug_alt
}
\keyword{dataset}

