\name{getStrata}
\alias{getStrata}
\title{Returns a data frame with two columns corresponding to the level 2 variable...}
\usage{getStrata(mlCtree)}
\description{Returns a data frame with two columns corresponding to the level 2 variable
and the leaves from the conditional inference trees.}
\value{a data frame}
