% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlpsa.print.R
\name{is.mlpsa}
\alias{is.mlpsa}
\title{Returns true if the object is of type \code{mlpsa}}
\usage{
is.mlpsa(x)
}
\arguments{
\item{x}{the object to test}
}
\description{
Returns true if the object is of type \code{mlpsa}
}

