\name{lsos}
\alias{lsos}
\title{Nicer list of objects in memory. Particularly useful for analysis of large data.
\url{# http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}}
\usage{
  lsos(..., n = 10)
}
\arguments{
  \item{n}{the number of objects to return.}

  \item{...}{not used.}
}
\value{
  a list of objects loaded sorted by size.
}
\description{
  Nicer list of objects in memory. Particularly useful for
  analysis of large data. \url{#
  http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}
}

