% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.missing.R
\name{missing.plot}
\alias{missing.plot}
\title{Returns a heat map graphic representing missingness of variables grouped by
the given grouping vector.}
\usage{
missing.plot(
  x,
  grouping,
  grid = FALSE,
  widths = c(ggplot2::unit(3, "null"), ggplot2::unit(1, "inches")),
  heights = c(ggplot2::unit(1, "inches"), ggplot2::unit(3, "null")),
  color = "red",
  ...
)
}
\arguments{
\item{x}{a data frame containing the variables to visualize missingness}

\item{grouping}{a vector of length nrow(vars) corresponding to how missing will be grouped by}

\item{grid}{whether to draw a grid between tiles}

\item{widths}{the ratio of the widths of the heatmap and histogram.}

\item{heights}{the ratio of the heights of the heatmap and histogram.}

\item{color}{the color used for indicating missingness.}

\item{...}{currently unused.}
}
\value{
a ggplot2 expression
}
\description{
NOTE: This is an experimental function and the results may vary depending on the
      nature of the dataset.
}
\seealso{
plot.mlpsa
}
