\name{multilevelCtree}
\alias{multilevelCtree}
\title{Performs a recursive partitioning in a conditional inference framework for each level 2.}
\usage{
  multilevelCtree(vars, formula, level2)
}
\value{
  a list of BinaryTree-class classes for each level 2
}
\description{
  TODO: Need more details
}

