\name{multilevelLR}
\alias{multilevelLR}
\title{Performs a logistic regression analysis for each level 2.}
\usage{multilevelLR(vars, formula, level2)}
\description{Performs a logistic regression analysis for each level 2.}
\details{TODO: Need more details}
\value{a list of glm classes for each level 2}
