\docType{package}
\name{multilevelPSA-package}
\alias{multilevelPSA}
\alias{multilevelPSA-package}
\title{Multilevel Propensity Score Analysis}
\description{
  This packages provides functions to estimate and
  visualize multilvel propensity score analys.
}
\details{
  This package extends the principles put forth by the
  \code{PSAgraphics} (Helmreich, Pruzek, & Xiong, 2010) for
  multilevel, or clustered, data.
}
\author{
  Jason Bryer \email{jason@bryer.org}
}
\references{
  \url{http://cran.r-project.org/web/packages/PSAgraphics/PSAgraphics.pdf}
  \url{http://www.jstatsoft.org/v29/i06/}
}
\seealso{
  \code{\link{PSAgraphics}}
}
\keyword{multilevel}
\keyword{psa}

