% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelPSA-package.R
\docType{data}
\name{pisa.countries}
\alias{pisa.countries}
\title{Data frame mapping PISA countries to their three letter abbreviation.}
\format{
data frame with 65 rows of 2 variables.
}
\description{
This data frame has two columns, \code{CNT3} for the three letter abbreviation
of each country and \code{Country} that provides the full country name in English.
}
\keyword{datasets}
