\name{pisa}
\alias{pisa}
\title{This function will load the 2009 PISA student data file. The data file is not
included with the package, but instead is downloaded upon first use into the
the package's installed data directory. Therefore, subsequent calls will simply
load the local version.}
\usage{
  pisa()
}
\description{
  This function will load the 2009 PISA student data file.
  The data file is not included with the package, but
  instead is downloaded upon first use into the the
  package's installed data directory. Therefore, subsequent
  calls will simply load the local version.
}
\references{
  Organisation for Economic Co-operation and Development
  (2009).  Programme for International Student Assessment
  (PISA).  \url{http://www.pisa.oecd.org/}
}
\seealso{
  \link{pisaschool}
}

