\docType{data}
\name{pisa.student}
\alias{pisa.student}
\title{Student results from the 2009 Programme of International Student Assessment (PISA)
as provided by the Organization for Economic Co-operation and Development (OECD).
See \url{http://www.pisa.oecd.org/} for more information including the code book.}
\format{a data frame with 475,460 ovservations of 305 variables from North America.}
\source{
  Organization for Economic Co-operation and Development
}
\description{
  Student results from the 2009 Programme of International
  Student Assessment (PISA) as provided by the Organization
  for Economic Co-operation and Development (OECD). See
  \url{http://www.pisa.oecd.org/} for more information
  including the code book.
}
\keyword{datasets}

