\docType{data}
\name{pisana}
\alias{pisana}
\title{North American (i.e. Canada, Mexico, and United States) student results of the 2009
Programm of International Student Assessment.}
\format{a data frame with 475,460 ovservations of 305 variables.}
\source{
  Organization for Economic Co-operation and Development
}
\description{
  Student results from the 2009 Programme of International
  Student Assessment (PISA) as provided by the Organization
  for Economic Co-operation and Development (OECD). See
  \url{http://www.pisa.oecd.org/} for more information
  including the code book.
}
\references{
  Organisation for Economic Co-operation and Development
  (2009).  Programme for International Student Assessment
  (PISA).  \url{http://www.pisa.oecd.org/}
}
\keyword{datasets}

