\docType{data}
\name{pisanaschool}
\alias{pisanaschool}
\title{North American (i.e. Canada, Mexico, and United States) school results of the 2009
Programm of International Student Assessment.}
\format{a data frame with 17,145 ovservations of 247 variables.}
\source{
  Organization for Economic Co-operation and Development
}
\description{
  School results from the 2009 Programme of International
  Student Assessment (PISA) as provided by the Organization
  for Economic Co-operation and Development (OECD). See
  \url{http://www.pisa.oecd.org/} for more information
  including the code book.
}
\references{
  Organisation for Economic Co-operation and Development
  (2009).  Programme for International Student Assessment
  (PISA).  \url{http://www.pisa.oecd.org/}
}
\keyword{dataset}
\keyword{datasets}

