% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.balance.R
\name{print.covariate.balance}
\alias{print.covariate.balance}
\title{Prints the overall effects before and after propensity score adjustment.}
\usage{
\method{print}{covariate.balance}(x, ...)
}
\arguments{
\item{x}{results of [covariate.balance()].}

\item{...}{unused.}
}
\description{
Prints the overall effects before and after propensity score adjustment.
}
