% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlpsa.print.R
\name{print.mlpsa}
\alias{print.mlpsa}
\title{Prints basic information about a \code{mlpsa} class.}
\usage{
\method{print}{mlpsa}(x, ...)
}
\arguments{
\item{x}{the \code{mlpsa} class.}

\item{...}{unused.}
}
\description{
Prints basic information about a \code{mlpsa} class.
}

