% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrange.r
\name{print.psrange}
\alias{print.psrange}
\title{Prints information about a psrange result.}
\usage{
\method{print}{psrange}(x, ...)
}
\arguments{
\item{x}{psrange to print info about.}

\item{...}{currently unused}
}
\description{
Prints information about a psrange result.
}
