% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrange.r
\name{summary.psrange}
\alias{summary.psrange}
\title{Prints the summary results of psrange.}
\usage{
\method{summary}{psrange}(object, ...)
}
\arguments{
\item{object}{psrange to print summary of.}

\item{...}{currently unused.}
}
\description{
Prints the summary results of psrange.
}

