\name{xtable.mlpsa}
\alias{xtable.mlpsa}
\title{Prints the results of \code{\link{mlpsa}} as a LaTeX table.}
\usage{
  xtable.mlpsa(x, digits = 2, include.note = TRUE, ...)
}
\arguments{
  \item{x}{results of \code{\link{mlpsa}}}

  \item{digits}{number of digits to print.}

  \item{include.note}{include a table note indicating how
  many rows were removed due to insufficient data within a
  strata.}
}
\description{
  Prints the results of \code{\link{mlpsa}} as a LaTeX
  table.
}

