% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.balance.R
\name{covariate.balance}
\alias{covariate.balance}
\title{Estimate covariate effect sizes before and after propensity score adjustment.}
\usage{
covariate.balance(covariates, treatment, level2, strata, abs = TRUE)
}
\arguments{
\item{covariates}{frame or matrix of covariates.}

\item{treatment}{vector of treatment indicators.}

\item{level2}{vector indicating level 2 membership.}

\item{strata}{strata indicators.}

\item{abs}{if TRUE absolute values of effect sizes will be plotted.}
}
\description{
Estimate covariate effect sizes before and after propensity score adjustment.
}
