% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.difftable.R
\name{difftable.plot}
\alias{difftable.plot}
\title{This function produces a ggplot2 figure containing the mean differences for
each level two, or cluster.}
\usage{
difftable.plot(x, fill.colors = NULL, legendlab = NULL, ...)
}
\arguments{
\item{x}{the results of [mlpsa()].}

\item{fill.colors}{the colors to use for each level two.}

\item{legendlab}{the label to use for the legend, or NULL to exclude.}

\item{...}{currently unused.}
}
\value{
a ggplot2 figure
}
\description{
This function produces a ggplot2 figure containing the mean differences for
each level two, or cluster.
}
