\name{geom_rug_alt}
\alias{geom_rug_alt}
\title{Alternate Marginal Rug Plot Geom for ggplot2}
\usage{
  geom_rug_alt(mapping = NULL, data = NULL,
    stat = "identity", position = "identity", ...)
}
\arguments{
  \item{mapping}{mapping between variables and aesthetics
  generated by aes}

  \item{data}{dataset used in this layer, if not specified
  uses plot dataset}

  \item{stat}{statistic used by this laye}

  \item{position}{position adjustment used by this layer}

  \item{...}{ignored}
}
\description{
  This alternate rug plot geom simply plots the rug tufts
  on the top and right (instead of the bottom and left)
}
\author{
  Hadley Wickham \url{http://had.co.nz}\cr

  With contributions by:\cr William E. J. Doane
  \email{wil@drdoane.com}\cr Brian A. Danielak
  \email{brian@briandk.com}
}
\references{
  Wickham, H. (2009). Ggplot2: Elegant Graphics for Data
  Analysis. New York: Springer.

  Wilkinson, L. (1999). The Grammar of Graphics. Statistics
  and computing. New York: Springer.
}
\seealso{
  \code{\link{geom_rug}}
}

