% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPropensityScores}
\alias{getPropensityScores}
\title{Returns a data frame with two columns corresponding to the level 2 variable
and the fitted value from the logistic regression.}
\usage{
getPropensityScores(lr.results, nStrata = 5)
}
\arguments{
\item{lr.results}{the results of \code{\link{mlpsa.logistic}}}

\item{nStrata}{number of strata within each level.}
}
\value{
a data frame
}
\description{
Returns a data frame with two columns corresponding to the level 2 variable
and the fitted value from the logistic regression.
}
\seealso{
mlpsa.logistic
}

