% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlpsa.ctree.R
\name{getStrata}
\alias{getStrata}
\title{Returns a data frame with two columns corresponding to the level 2 variable
and the leaves from the conditional inference trees.}
\usage{
getStrata(party.results, data, level2)
}
\arguments{
\item{party.results}{the results of \code{\link{mlpsa.ctree}}}

\item{data}{the data frame to merge results to}

\item{level2}{the name of the level 2 variable.}
}
\value{
a data frame
}
\description{
Returns a data frame with two columns corresponding to the level 2 variable
and the leaves from the conditional inference trees.
}
\seealso{
mlpsa.ctree
}

