% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lsos}
\alias{lsos}
\title{Nicer list of objects in memory. Particularly useful for analysis of large data.
\url{# http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}}
\usage{
lsos(..., n = 10)
}
\arguments{
\item{...}{not used.}

\item{n}{the number of objects to return.}
}
\value{
a list of objects loaded sorted by size.
}
\description{
Nicer list of objects in memory. Particularly useful for analysis of large data.
\url{# http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}
}

