\docType{data}
\name{pisa.psa.cols}
\alias{pisa.psa.cols}
\title{Character vector representing the list of covariates used for estimating
propensity scores.}
\format{a character vector with covariate names for estimating propensity scores.}
\description{
  Character vector representing the list of covariates used
  for estimating propensity scores.
}
\keyword{datasets}

