\docType{data}
\name{pisana}
\alias{pisana}
\title{North American (i.e. Canada, Mexico, and United States) student results of the 2009
Programm of International Student Assessment.}
\format{a data frame with 66,548 ovservations of 65 variables.}
\source{
  Organization for Economic Co-operation and Development
}
\description{
  Student results from the 2009 Programme of International
  Student Assessment (PISA) as provided by the Organization
  for Economic Co-operation and Development (OECD). See
  \url{http://www.pisa.oecd.org/} for more information
  including the code book.
}
\details{
  Note that missing values have been imputed using the
  \href{mice}{http://cran.r-project.org/web/packages/mice/index.html}
  package. Details on the specific procedure are in the
  \code{pisa.impute} function in the
  \href{http://github.com/jbryer/pisa}{\code{pisa}
  package}.
}
\references{
  Organisation for Economic Co-operation and Development
  (2009).  Programme for International Student Assessment
  (PISA).  \url{http://www.pisa.oecd.org/}
}
\keyword{datasets}

