% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelPSA-package.R
\docType{data}
\name{pisana}
\alias{pisana}
\title{North American (i.e. Canada, Mexico, and United States) student results of the 2009
Programme of International Student Assessment.}
\format{
a data frame with 66,548 obvservations of 65 variables.
}
\source{
Organization for Economic Co-operation and Development
}
\description{
Student results from the 2009 Programme of International Student Assessment (PISA)
as provided by the Organization for Economic Co-operation and Development (OECD).
See [https://www.oecd.org/en/about/programmes/pisa.html/](https://www.oecd.org/en/about/programmes/pisa.html/)
for more information including the code book.
}
\details{
Note that missing values have been imputed using the 
[mice](https://cran.r-project.org/package=mice) package.
Details on the specific procedure are in the `pisa.impute` function
in the [pisa](https://github.com/jbryer/pisa) package.
}
\references{
Organization for Economic Co-operation and Development (2009).
            Programme for International Student Assessment (PISA). 
            [https://www.oecd.org/en/about/programmes/pisa.html](https://www.oecd.org/en/about/programmes/pisa.html)
}
\keyword{datasets}
