\name{plot.mlpsa}
\alias{plot.mlpsa}
\title{Plots the results of a multilevel propensity score model.}
\usage{
  \method{plot}{mlpsa} (mlpsa, ...)
}
\arguments{
  \item{mlpsa}{}
}
\description{
  The plot created uses the \code{ggplot2} framework. As
  such, additional modificaitons can be made. This plot is
  an extension of the \code{circ.psa} function in the
  \code{PSAgraphics} package for multilevel models.
}

