% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.covariate.balance}
\alias{print.covariate.balance}
\title{Prints the overall effects before and after propensity score adjustment.}
\usage{
\method{print}{covariate.balance}(x, ...)
}
\arguments{
\item{x}{results of \code{\link{covariate.balance}}.}

\item{...}{unused.}
}
\description{
Prints the overall effects before and after propensity score adjustment.
}

