\name{print.xmlpsa}
\alias{print.xmlpsa}
\title{Prints the results of \code{\link{mlpsa}} and \code{\link{xtable.mlpsa}}.}
\usage{
  \method{print}{xmlpsa} (x,
    tabular.environment = "longtable", floating = FALSE,
    ...)
}
\arguments{
  \item{x}{result of \code{\link{xtable.mlpsa}}}

  \item{tabular.environment}{see
  \code{\link{print.xtable}}.}

  \item{floating}{see \code{\link{print.xtable}}.}

  \item{...}{other parameters passed to
  \code{\link{print.xtable}}}
}
\description{
  Print method for \code{\link{xtable.mlpsa}}.
}

